Moving plan modifiers to a separate file (#177)
