+++ internal/provider/string.go
@@ -154,87 +154,6 @@ func getStringSchemaV1(sensitive bool, description string) tfsdk.Schema {
 	}
 }
 
-//nolint:unparam
-func defaultBool(val bool) tfsdk.AttributePlanModifier {
-	return boolDefault{val}
-}
-
-type boolDefault struct {
-	val bool
-}
-
-func (d boolDefault) Description(ctx context.Context) string {
-	return "If the plan does not contain a value, a default will be set using val."
-}
-
-func (d boolDefault) MarkdownDescription(ctx context.Context) string {
-	return "If the plan does not contain a value, a default will be set using `val`."
-}
-
-func (d boolDefault) Modify(ctx context.Context, req tfsdk.ModifyAttributePlanRequest, resp *tfsdk.ModifyAttributePlanResponse) {
-	t := req.AttributeConfig.(types.Bool)
-
-	if t.Null {
-		resp.AttributePlan = types.Bool{
-			Value: d.val,
-		}
-	}
-}
-
-func defaultInt(val int64) tfsdk.AttributePlanModifier {
-	return intDefault{val}
-}
-
-type intDefault struct {
-	val int64
-}
-
-func (d intDefault) Description(ctx context.Context) string {
-	return "If the plan does not contain a value, a default will be set using val."
-}
-
-func (d intDefault) MarkdownDescription(ctx context.Context) string {
-	return "If the plan does not contain a value, a default will be set using `val`."
-}
-
-func (d intDefault) Modify(ctx context.Context, req tfsdk.ModifyAttributePlanRequest, resp *tfsdk.ModifyAttributePlanResponse) {
-	t := req.AttributeConfig.(types.Int64)
-
-	if t.Null {
-		resp.AttributePlan = types.Int64{
-			Null:  false,
-			Value: d.val,
-		}
-	}
-}
-
-func defaultString(val string) tfsdk.AttributePlanModifier {
-	return stringDefault{val}
-}
-
-type stringDefault struct {
-	val string
-}
-
-func (d stringDefault) Description(ctx context.Context) string {
-	return "If the plan does not contain a value, a default will be set."
-}
-
-func (d stringDefault) MarkdownDescription(ctx context.Context) string {
-	return "If the plan does not contain a value, a default will be set."
-}
-
-func (d stringDefault) Modify(ctx context.Context, req tfsdk.ModifyAttributePlanRequest, resp *tfsdk.ModifyAttributePlanResponse) {
-	t := req.AttributeConfig.(types.String)
-
-	if t.Null {
-		resp.AttributePlan = types.String{
-			Null:  false,
-			Value: d.val,
-		}
-	}
-}
-
 func createString(ctx context.Context, req tfsdk.CreateResourceRequest, resp *tfsdk.CreateResourceResponse, sensitive bool) {
 	const numChars = "0123456789"
 	const lowerChars = "abcdefghijklmnopqrstuvwxyz"
